支付修改

+++ car/src/main/java/cc/carce/sale/dto/CarShoppingCartDto.java
@@ -33,7 +33,7 @@ public class CarShoppingCartDto {
     /**
      * 产品价格
      */
-    private Integer productPrice;
+    private Long productPrice;
     
     /**
      * 产品名称
